* 'master' of ssh://mcminos.com:80/mcminos-mobile:
  don't checkin robovm.prperties anybore as it's generated
  change robovm properties automatically
  also ignore the flavor-config
  some files forgotten to delte and ignore
  adding flavors for game icons
  reorganization for flavors and adding relevant scripts
  cleaning up gradle files
  add crash analytics and obfuscation
  removed something is odd from alive - added close quarters
  fixed music in credits

+++ tools/set_version.sh
@@ -1,5 +1,14 @@
 #!/bin/bash
 export LANG=C
+
+dir=$(cd -P -- "$(dirname -- "$0")" && pwd -P)
+cd "$dir/.."
+
+if test ! -e FLAVOR; then
+  echo "Fatal Error: Make sure you have selected a flavor (run tools/flavor <flavor>)."
+  exit 1
+fi
+
 DESTFOLDER="android/assets"
 revision=$(git rev-parse --short HEAD)
 cat > "$DESTFOLDER/VERSION" << EOF
